+++ TinyBunnyJump/Source/QQSpriteBalloonBlinker.mm
@@ -66,7 +66,7 @@
             [self removeSelf];
             
             
-            int scoreMultiplier = 1;
+            float scoreMultiplier = 1;
             QQBalloonClass* myScoreInfo = (QQBalloonClass*)[self userInfo];
             if(myScoreInfo != nil){
                 scoreMultiplier = [myScoreInfo scoreMultiplier];
